requirements.txt
absl-py==2.3.1
accelerate==1.10.0
addict==2.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
diffusers==0.35.0
filelock==3.13.1
fonttools==4.59.1
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.74.0
h5py==3.14.0
hf-xet==1.1.8
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.9
lightning-utilities==0.15.2
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.5
mdurl==0.1.2
mmengine==0.10.7
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.0.0
platformdirs==4.3.8
protobuf==6.32.0
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
sentry-sdk==2.35.0
six==1.17.0
smmap==5.0.2
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.21.4
tomli==2.2.1
torch==2.4.1+cu121
torchinfo==1.8.0
torchmetrics==1.8.1
torchvision==0.19.1+cu121
tqdm==4.67.1
transformers==4.55.2
triton==3.0.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.1
Werkzeug==3.1.3
xformers==0.0.28
yapf==0.43.0
zipp==3.23.0
